Move missed tests from tests/ directory
